changed the way atlas is integrated into forward rendering and clusters
